게시글 사이드바 안보이는 버그 수정

+++ src/pages/BoardPage/index.tsx
@@ -30,6 +30,7 @@ const PostPage = () => {
 
   useEffect(() => {
     checkLike();
+    idTag();
   }, [board]);
 
   const getBoard = async () => {
@@ -81,6 +82,8 @@ const PostPage = () => {
   };
 
   const delBoard = async () => {
+    if (confirm("정말 삭제하시겠습니까?") === false) return;
+
     await deleteBoard(Number(boardId), user.id)
       .then(() => {
         navigate("/");
@@ -219,7 +222,7 @@ const PostPage = () => {
 
             <SideContainer>
               <SideWrapper>
-                <SideTool fixed={fixed ? "true" : "false"}>
+                <SideTool $fixed={fixed}>
                   <Favorite
                     onClick={handleLike}
                     color={like ? "error" : "disabled"}
@@ -240,7 +243,7 @@ const PostPage = () => {
 
             <SideContainer>
               <SideNavWrapper>
-                <SideNav fixed={fixed ? "true" : "false"}>
+                <SideNav $fixed={fixed}>
                   {headerArray.map((item, index) => (
                     <SideNavTitle key={index}>
                       <Link
@@ -357,8 +360,8 @@ const SideNavWrapper = styled.div`
   position: absolute;
   left: 100%;
 `;
-const SideNav = styled.div<{ fixed: string }>`
-  position: ${(props) => (props.fixed === "true" ? "fixed" : "relative")};
+const SideNav = styled.div<{ $fixed: boolean }>`
+  position: ${({ $fixed }) => ($fixed ? "fixed" : "absolute")};
   top: 122px;
   width: 240px;
   max-height: calc(100vh - 128px);
@@ -379,7 +382,8 @@ const SideNavTitle = styled.div`
   transition: all 0.125s ease-in 0s;
 
   a.active {
-    transform: scale(1.1);
+    transform: scale(1.2);
+    font-weight: 600;
     color: #212529;
   }
 
@@ -388,8 +392,8 @@ const SideNavTitle = styled.div`
   }
 `;
 
-const SideTool = styled.div<{ fixed: string }>`
-  position: ${(props) => (props.fixed === "true" ? "fixed" : "absolute")};
+const SideTool = styled.div<{ $fixed: boolean }>`
+  position: ${({ $fixed }) => ($fixed ? "fixed" : "absolute")};
   background-color: #f8f9fa;
   top: 122px;
   display: flex;
